+++ assets/src/App.css
@@ -50,13 +50,13 @@ body {
   color: #fff;
 }
 
-a.RelatedCustomerConversations {
+a.RelatedCustomerConversation--link {
   display: block;
   color: rgba(0, 0, 0, 0.45);
   transition: all 0.3s cubic-bezier(0.645, 0.045, 0.355, 1);
 }
 
-a.RelatedCustomerConversations:hover {
+a.RelatedCustomerConversation--link:hover {
   background: rgba(0, 0, 0, 0.018);
 }
